g4-common: Add AuroraStore alternative to PlayStore
- After a through out checking in our props i couldn't find the reason
behind play store error "Device is not compatible".
As a matter of fact we use the same exactly props in other roms too
like LineageOS and others, but in every other rom we don't have this issue.

Since i am unable to find the prop that causes the error, use an alternative
store to help us find, install and update our apps.

- Can be used alongside with PlayStore or if user wants he can disable PlayStore

Change-Id: Ib95dadd3a1ad3d12d56d8d01f4fa409929150adc